fix database issue

+++ database/migrations/2024_01_27_083127_create_payment_methods_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('payment_methods', function (Blueprint $table) {
             $table->id();
             $table->string('payment_method_name');
-            $table->double('service_charge');
+            $table->integer('service_charge');
             $table->boolean('isActive');
             $table->timestamps();
         });
